feat(ui): WhisperModelPicker partage + gestionnaire de modeles dans Reglages (spec/04/11)
- nouveau composant autonome : liste list_whisper_models, etats
  telecharge (badge Actif / Utiliser / supprimer) / en telechargement
  (barre + % + Annuler) / absent (Telecharger, Reessayer sur erreur) ;
  suit les events par nom de modele -> re-affiche une progression deja
  en cours ; premier modele telecharge devient actif
- Settings : la section Transcription remplace le select legacy + listener
  download-progress par le picker (fin du footgun set_config avant un
  telechargement qui peut echouer) ; ligne Langue inchangee
- size_mb passe en u32 (ts-rs: number, pas bigint)

+++ src-tauri/src/transcription/mod.rs
@@ -17,7 +17,7 @@ use uuid::Uuid;
 /// ggerganov/whisper.cpp) ; `size_mb` ne sert qu'à l'affichage.
 pub struct WhisperModelDef {
     pub name: &'static str,
-    pub size_mb: u64,
+    pub size_mb: u32,
     pub recommended: bool,
 }
 
@@ -37,7 +37,7 @@ pub type DownloadRegistry = Arc<StdMutex<HashMap<String, Arc<AtomicBool>>>>;
 #[ts(export, export_to = "../../src/bindings/")]
 pub struct WhisperModelInfo {
     pub name: String,
-    pub size_mb: u64,
+    pub size_mb: u32,
     pub recommended: bool,
     /// "downloaded" | "downloading" | "missing"
     pub status: String,
